requirements.txt
click==8.1.7
markdown-it-py==3.0.0
mdurl==0.1.2
Pygments==2.17.2
python-dotenv==1.0.1
rich==13.7.1
